Revert "API-47340 - Update v1 PDF uploads to use Firefly service (#24381)"
This reverts commit 246aabed02a9a110c9a53b7277b6fc3485c47fd1.
