+++ modules/claims_api/app/controllers/claims_api/v2/veterans/disability_compensation_controller.rb
@@ -232,7 +232,7 @@ module ClaimsApi
         end
 
         def bd_service
-          ClaimsApi::DisabilityCompensationBenefitsDocumentsUploader
+          ClaimsApi::V2::DisabilityCompensationBenefitsDocumentsUploader
         end
 
         def sandbox_request(request)
